global doc
