+++ plugins/chat/lib/tasks/chat_doc.rake
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 task "chat:doc" do
-  ["plugins/chat/assets/javascripts/discourse/services/chat-api.js"].each do |file|
-    origin = File.join(Rails.root, file)
-    filename = File.basename(file, ".*")
-    destination = File.join(Rails.root, "plugins/chat/docs/#{filename}.md")
-    config = File.join(Rails.root, ".jsdoc")
+  destination = File.join(Rails.root, "plugins/chat/docs/FRONTEND.md")
+  config = File.join(Rails.root, ".jsdoc")
 
-    `jsdoc2md -c #{config} #{origin} > #{destination}`
-  end
+  files = %w[
+    plugins/chat/assets/javascripts/discourse/lib/collection.js
+    plugins/chat/assets/javascripts/discourse/services/chat-api.js
+  ]
+
+  `jsdoc2md --separators -c #{config} -f #{files.join(" ")} > #{destination}`
 end
